+++ src/Resources/app/administration/src/module/postfinancecheckout-settings/page/postfinancecheckout-settings/configuration-constants.js
@@ -8,6 +8,8 @@ export const CONFIG_SPACE_VIEW_ID = CONFIG_DOMAIN + '.' + 'spaceViewId';
 export const CONFIG_STOREFRONT_INVOICE_DOWNLOAD_ENABLED = CONFIG_DOMAIN + '.' + 'storefrontInvoiceDownloadEnabled';
 export const CONFIG_USER_ID = CONFIG_DOMAIN + '.' + 'userId';
 export const CONFIG_IS_SHOWCASE = CONFIG_DOMAIN + '.' + 'isShowcase';
+export const CONFIG_STOREFRONT_WEBHOOKS_UPDATE_ENABLED = CONFIG_DOMAIN + '.' + 'storefrontWebhooksUpdateEnabled';
+export const CONFIG_STOREFRONT_PAYMENTS_UPDATE_ENABLED = CONFIG_DOMAIN + '.' + 'storefrontPaymentsUpdateEnabled';
 
 export default {
     CONFIG_DOMAIN,
@@ -19,5 +21,7 @@ export default {
     CONFIG_SPACE_VIEW_ID,
     CONFIG_STOREFRONT_INVOICE_DOWNLOAD_ENABLED,
     CONFIG_USER_ID,
-    CONFIG_IS_SHOWCASE
+    CONFIG_IS_SHOWCASE,
+    CONFIG_STOREFRONT_WEBHOOKS_UPDATE_ENABLED,
+    CONFIG_STOREFRONT_PAYMENTS_UPDATE_ENABLED
 };